--- New Unity Project/Assets/code/Player/Test 1 old.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b36ec544253f6d4baeb7ba145311acd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
