--- New Unity Project/Assets/code/Player/Test 1 old/PlayerMovement.cs
@@ -1,26 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-//public class PlayerMovement : MonoBehaviour
-
-//{
-
-//    public CharacterController2D controller;
-//    
-//   [SerializeField]
-//    private float runSpeed = 4f;
-//    
-
-//    void Update()
-//    {
-//        float horizontalMove = Input.GetAxisRaw("Horizontal") * runSpeed;
-//        Vector3 direction = new Vector3 (horizontalMove, 0, 0);
-//        transform.Translate(direction * runSpeed * Time.deltaTime);
-//    }
-
-//    void FixedUpdate() 
-//    {
-//        // controller.Move(horizontalMove * Time.fixedDeltaTime, false, false);
-//    }
-//}
