--- New Unity Project/Assets/code/Player/Test 1 old/PlayerMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 689b8f4dad9082249b088a526e5a9b65
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
