+++ New Unity Project/Assets/code/Player/Test 1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b36ec544253f6d4baeb7ba145311acd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
