+++ New Unity Project/Assets/code/Player/Test 1/CharacterController2D.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2367424c69d3ab048adac8b447d02069
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
