--- New Unity Project/Assets/code/Player/Test 2 working/Character2DController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 489b93c3914cf384288eb16ed8aa4588
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
