+++ New Unity Project/Assets/code/Player/Test 2.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a8473553e522f9468ba9fa282ddff7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
