+++ New Unity Project/Library/StateCache/SceneView/85b21c-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":36.840450286865237,"y":2.9812984466552736,"z":-20.5078125},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":23.610586166381837,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":1.6637277603149415,"y":-0.22936324775218965,"z":-12.59333610534668},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":4.404995441436768,"orthographic":true}
